+++ app/models/game_item_category.rb
@@ -1,0 +1,4 @@
+# Represents game item categories
+class GameItemCategory < ApplicationRecord
+  has_many :game_items, dependent: :destroy
+end