+++ app/models/game_item_subcategory.rb
@@ -1,0 +1,4 @@
+# Represents game item subcategories
+class GameItemSubcategory < ApplicationRecord
+  has_many :game_items, dependent: :destroy
+end