+++ spec/factories/game_item_categories.rb
@@ -1,0 +1,6 @@
+FactoryBot.define do
+  factory :game_item_category do
+    uid { 7 }
+    name { { 'en' => 'Trade Goods', 'ru' => 'Хозяйственные товары' } }
+  end
+end