+++ spec/factories/game_item_subcategories.rb
@@ -1,0 +1,6 @@
+FactoryBot.define do
+  factory :game_item_subcategory do
+    uid { 7 }
+    name { { 'en' => 'Metal & Stone', 'ru' => 'Металл и камни' } }
+  end
+end