+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - api endpoint for import bank data
 - BankCell model
 - creating/updating/deleting bank cells from addon data
+- GameItem models
 
 ## [1.0.1] - 2019-12-11
 ### Added